# docs/openapi-gpt-3.0.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "esdata API",
    "description": "API de consulta fiscal-regulatoria con trazabilidad a fuente oficial, superficies de governance y workers observables via sync_log.",
    "version": "0.1.6"
  },
  "servers": [
    {
      "url": "https://api.desuscribir.es",
      "description": "Production API"
    }
  ],
  "paths": {
...
    "/v1/internacional/obligaciones": {
      "get": {
        "tags": [
          "internacional"
        ],
        "summary": "Listar Obligaciones Internacionales",
        "operationId": "listar_obligaciones_internacionales",
        "parameters": [
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por tipo: tratado, convenio, directiva, ley",
              "title": "Tipo"
            },
            "description": "Filtrar por tipo: tratado, convenio, directiva, ley"
          },
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Estado: activo, inactivo, obsoleto",
              "default": "activo",
              "title": "Estado"
            },
            "description": "Estado: activo, inactivo, obsoleto"
          },
          {
            "name": "jurisdiccion",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Filtrar por jurisdiccion de origen",
              "title": "Jurisdiccion"
            },
            "description": "Filtrar por jurisdiccion de origen"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "description": "Numero maximo de registros",
              "default": 100,
              "title": "Limit"
            },
            "description": "Numero maximo de registros"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Desplazamiento para paginacion",
              "default": 0,
              "title": "Offset"
            },
            "description": "Desplazamiento para paginacion"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligacionInternacionalListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/internacional/obligaciones/{codigo}": {
      "get": {
        "tags": [
          "internacional"
        ],
        "summary": "Detalle Obligacion Internacional",
        "operationId": "detalle_obligacion_internacional",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Codigo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObligacionInternacionalDetailResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "ObligacionInternacionalDetailResponse": {
        "properties": {
          "item": {
            "$ref": "#/components/schemas/ObligacionInternacionalItem"
          }
        },
        "type": "object",
        "required": [
          "item"
        ],
        "title": "ObligacionInternacionalDetailResponse"
      },
      "ObligacionInternacionalItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "codigo": {
            "type": "string",
            "title": "Codigo"
          },
          "titulo": {
            "type": "string",
            "title": "Titulo"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "jurisdiccion_origen": {
            "type": "string",
            "title": "Jurisdiccion Origen"
          },
          "jurisdiccion_aplicacion": {
            "type": "string",
            "title": "Jurisdiccion Aplicacion"
          },
          "vigente_desde": {
            "type": "string",
            "title": "Vigente Desde"
          },
          "vigente_hasta": {
            "type": "string",
            "title": "Vigente Hasta"
          },
          "estado": {
            "type": "string",
            "title": "Estado"
          },
          "descripcion": {
            "type": "string",
            "title": "Descripcion"
          },
          "creado_en": {
            "type": "string",
            "title": "Creado En"
          },
          "actualizado_en": {
            "type": "string",
            "title": "Actualizado En"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url",
            "description": "URL oficial registrada en source_revision"
          },
          "source_worker": {
            "type": "string",
            "title": "Source Worker",
            "description": "Worker que cargo la referencia"
          },
          "source_fetched_at": {
            "type": "string",
            "title": "Source Fetched At",
            "description": "Fecha de captura de la fuente"
          }
        },
        "type": "object",
        "required": [
          "id",
          "codigo",
          "titulo",
          "tipo",
          "estado"
        ],
        "title": "ObligacionInternacionalItem"
      },
      "ObligacionInternacionalListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ObligacionInternacionalItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_offset": {
            "type": "integer",
            "title": "Next Offset"
          }
        },
        "type": "object",
        "required": [
          "total"
        ],
        "title": "ObligacionInternacionalListResponse"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Dedicated ESData API key for GPT Actions."
      }
    }
  }
}